+++ src/main/java/org/ever/_4ever_be_gw/business/dto/analytics/TrendPointDto.java
@@ -5,15 +5,17 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class TrendPointDto {
-    private Integer year;
-    private Integer month;
-    private Integer week;
-    private Long sale;
-    private Integer orderCount;
+    private int year;
+    private int month;
+    private int week;
+    private BigDecimal sale;
+    private Long orderCount;
 }
